Move all point-related code to geometric package
