+++ src/main/java/org/wmn4j/representation/geometric/package-info.java
@@ -1,0 +1,7 @@
+/*
+ * Distributed under the MIT license (see LICENSE.txt or https://opensource.org/licenses/MIT).
+ */
+/**
+ * Contains geometric representations of music.
+ */
+package org.wmn4j.representation.geometric;
